+++ tests/test_utils.py
@@ -1,4 +1,5 @@
 import os
+import stat
 import sys
 
 import pytest
@@ -6,7 +7,7 @@ import pytest
 import click
 import click.utils
 import click._termui_impl
-from click._compat import WIN, PY2
+from click._compat import WIN, PY2, _get_current_umask
 
 
 def test_echo(runner):
@@ -283,6 +284,43 @@ def test_open_file(runner):
         result = runner.invoke(cli, ['-'], input='foobar')
         assert result.exception is None
         assert result.output == 'foobar\nmeep\n'
+
+
+@pytest.mark.skipif(WIN, reason='os.chmod() is not fully supported on Windows.')
+@pytest.mark.parametrize('permissions', [
+    0o400,
+    0o444,
+    0o600,
+    0o644,
+ ])
+def test_open_file_atomic_permissions_existing_file(runner, permissions):
+    with runner.isolated_filesystem():
+        with open('existing.txt', 'w') as f:
+            f.write('content')
+        os.chmod('existing.txt', permissions)
+
+        @click.command()
+        @click.argument('filename')
+        def cli(filename):
+           click.open_file(filename, 'w', atomic=True).close()
+
+        result = runner.invoke(cli, ['existing.txt'])
+        assert result.exception is None
+        assert stat.S_IMODE(os.stat('existing.txt').st_mode) == permissions
+
+
+@pytest.mark.skipif(WIN, reason='os.chmod() is not fully supported on Windows.')
+def test_open_file_atomic_permissions_new_file(runner):
+    with runner.isolated_filesystem():
+        @click.command()
+        @click.argument('filename')
+        def cli(filename):
+            click.open_file(filename, 'w', atomic=True).close()
+
+        umask = _get_current_umask()
+        result = runner.invoke(cli, ['new.txt'])
+        assert result.exception is None
+        assert stat.S_IMODE(os.stat('new.txt').st_mode) == (0o666 & ~umask)
 
 
 @pytest.mark.xfail(WIN and not PY2, reason='God knows ...')